feat: fractional quantities for perfume refills
DDL:
- widen quantity columns to DECIMAL(10,3): shop_stock, sale_items, transfer_items, purchase_items, products.stock_quantity, products.wholesale_min_qty, purchases.total_items, stock_transfers.total_items
- add products.unit_type ENUM('package','unit'), unit_label VARCHAR(20)
- add chk_unit_label CHECK constraint
- flag 97 Perfume Re-fill products (category_id 46, company 2) as unit/ml

Go:
- widen all quantity fields to float64 across db + handlers
- add FormatQty helper (trim trailing zeros)
- add unit_type/unit_label to Product struct + scan in GetProductByBarcodeAndShop
- fix all quantity format strings

Frontend:
- parseFloat on all cart quantity paths
- min/step 0.5 for unit products in RenderRowHTML
- includes uncommitted template/JS edits from prior session

Docs:
- migrations/20260923_FRACTIONAL_QTY_LOG.md (deploy plan, rollback, GAP list)
- PROJECT_CONTEXT.md (positional Scan rule, float64 convention)

Also untracks spide-pos binaries (added to .gitignore).

+++ internal/db/dashboard.go
@@ -168,13 +168,13 @@ func GetDashboardStats(db *sql.DB, shopID int) (*DashboardStats, error) {
 
 	topProducts := []struct {
 		ProductName string  `json:"product_name"`
-		UnitsSold   int     `json:"units_sold"`
+		UnitsSold   float64 `json:"units_sold"`
 		Revenue     float64 `json:"revenue"`
 	}{}
 	for rows.Next() {
 		var tp struct {
 			ProductName string  `json:"product_name"`
-			UnitsSold   int     `json:"units_sold"`
+			UnitsSold   float64 `json:"units_sold"`
 			Revenue     float64 `json:"revenue"`
 		}
 		if err := rows.Scan(&tp.ProductName, &tp.UnitsSold, &tp.Revenue); err != nil {
